+++ ci.nix
@@ -9,7 +9,6 @@
   inherit (import ./ci-lib.nix) dimension platformFilterGeneric filterAttrsOnlyRecursive;
   sources = import ./nix/sources.nix {};
   nixpkgsVersions = {
-    "R1909" = "nixpkgs-1909";
     "R2003" = "nixpkgs-2003";
     "R2009" = "nixpkgs-2009";
   };