+++ overlays/shims.nix
@@ -1,9 +1,0 @@
-# Shims to make older versions of nixpkgs work
-final: prev: {
-  # nixpkgs 19.09 and older does not have runCommandLocal
-  runCommandLocal = prev.runCommandLocal or (name: env:
-    final.runCommand name ({
-      preferLocalBuild = true;
-      allowSubstitutes = false;
-    } // env));
-}